requirements.txt
aio-pika==7.0.1
aiofiles==0.8.0
aiohttp==3.8.1
aiormq==6.2.3
aiosignal==1.2.0
amqp==5.1.1
anyio==3.5.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
base58==2.1.1
billiard==3.6.4.0
bitarray==1.2.2
black==22.3.0
celery==5.2.3
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.11.2
distlib==0.3.4
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
filelock==3.6.0
flake8==4.0.1
flower==1.0.0
frozenlist==1.3.0
greenlet==1.1.2
h11==0.12.0
hexbytes==0.2.2
httpcore==0.14.7
httpx==0.22.0
humanize==4.0.0
identify==2.5.0
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
kombu==5.2.4
loguru==0.6.0
lru-dict==1.1.7
mccabe==0.6.1
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
nodeenv==1.6.0
pamqp==3.1.0
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.18.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycryptodome==3.14.1
pyflakes==2.4.0
pyrsistent==0.18.1
PySocks==1.7.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
rlp==2.0.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
typing-extensions==4.2.0
urllib3==1.26.9
uvloop==0.16.0
varint==1.0.2
vine==5.0.0
virtualenv==20.14.1
wcwidth==0.2.5
web3==5.28.0
websockets==9.1
yarl==1.7.2
